+++ autogalaxy/__init__.py
@@ -20,7 +20,6 @@ from autoarray.inversion.mappers.factory import mapper_from as Mapper
 from autoarray.inversion.pixelizations.settings import SettingsPixelization
 from autoarray.mask.mask_1d import Mask1D
 from autoarray.mask.mask_2d import Mask2D
-from autoarray import fixtures
 from autoarray.operators.convolver import Convolver
 from autoarray.operators.convolver import Convolver
 from autoarray.operators.transformer import TransformerDFT
@@ -78,6 +77,7 @@ from .profiles import (
 from .profiles.light_profiles import light_profiles_init as lp_init
 from .profiles.light_profiles import light_profiles_snr as lp_snr
 from . import convert
+from . import mock as m
 from .util.shear_field import ShearYX2D
 from .util.shear_field import ShearYX2DIrregular
