+++ autogalaxy/mock.py
@@ -1,0 +1,7 @@
+from autofit.mock import  *
+from autoarray.mock import *
+
+from autogalaxy.analysis.mock.mock_result import MockResult
+from autogalaxy.profiles.light_profiles.mock.mock_light_profile import MockLightProfile
+from autogalaxy.profiles.mass_profiles.mock.mock_mass_profile import MockMassProfile
+from autogalaxy.util.mock.mock_cosmology import MockCosmology
