+++ autogalaxy/profiles/mass_profiles/mock/mock_mass_profile.py
@@ -1,0 +1,30 @@
+import autogalaxy as ag
+
+
+class MockMassProfile(ag.mp.MassProfile):
+    def __init__(
+        self,
+        convergence_2d=None,
+        potential_2d=None,
+        deflections_yx_2d=None,
+        value=None,
+        value1=None,
+    ):
+
+        super().__init__()
+
+        self.convergence_2d = convergence_2d
+        self.potential_2d = potential_2d
+        self.deflections_2d = deflections_yx_2d
+
+        self.value = value
+        self.value1 = value1
+
+    def convergence_2d_from(self, grid):
+        return self.convergence_2d
+
+    def potential_2d_from(self, grid):
+        return self.potential_2d
+
+    def deflections_yx_2d_from(self, grid):
+        return self.deflections_2d
