+++ test_autogalaxy/imaging/model/test_aggregator_imaging.py
@@ -7,7 +7,6 @@ from autoconf import conf
 import autofit as af
 import autogalaxy as ag
 from autofit.non_linear.samples import Sample
-from autogalaxy.mock import mock
 
 directory = path.dirname(path.realpath(__file__))
 
@@ -44,7 +43,7 @@ def make_samples(model):
         weight_list=[0.0, 1.0],
     )
 
-    return mock.MockSamples(
+    return ag.m.MockSamples(
         model=model, sample_list=sample_list, max_log_likelihood_instance=plane
     )
 
@@ -78,7 +77,7 @@ def test__imaging_generator_from_aggregator(imaging_7x7, mask_2d_7x7, samples, m
         )
     )
 
-    search = mock.MockSearch(samples=samples)
+    search = ag.m.MockSearch(samples=samples)
     search.paths = af.DirectoryPaths(path_prefix=path_prefix)
 
     analysis = ag.AnalysisImaging(dataset=masked_imaging_7x7)
