+++ test_autogalaxy/profiles/mass_profiles/test_total_mass_profiles.py
@@ -1,7 +1,7 @@
-import autogalaxy as ag
 import numpy as np
 import pytest
 
+import autogalaxy as ag
 
 grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
