+++ test_autogalaxy/quantity/test_dataset_quantity.py
@@ -3,7 +3,6 @@ import pytest
 
 import autogalaxy as ag
 from autogalaxy import exc
-from autogalaxy.mock import mock
 
 
 def test_via_signal_to_noise_map(dataset_quantity_7x7_array_2d, sub_mask_2d_7x7):
